requirements.txt
# 每日提醒助手 - 依赖包
# 基础版（GitHub Actions运行）

# HTTP请求（天气API、Server酱）
requests>=2.28.0

# Excel解析（课表导入）
pandas>=2.0.0
openpyxl>=3.1.0

# 环境变量（本地测试用）
python-dotenv>=1.0.0

# 时区处理（GitHub Actions 时区问题修复）
pytz>=2024.1

# Web配置界面
flask>=2.3.0
werkzeug>=2.3.0
